refactor(doctor): centralize checks in handler
