+++ src/handlers/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod deploy;
+pub(crate) mod doctor;
 pub(crate) mod evolve;
 pub(crate) mod read_only;
 pub(crate) mod rollback;
